requirements.txt
PyQt5==5.15.10
pytesseract==0.3.10
Pillow==10.2.0
deep-translator==1.11.4